src/ldr-sensor/requirements.txt
paho-mqtt==1.4.0
netifaces==0.10.9
